Moved quantity, size, and style selectors to separate component files. Size and quantity selectors behaving as desired.
